requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cycler==0.10.0
dataclasses==0.6
docker-pycreds==0.4.0
evdev==1.4.0
fsspec==2021.4.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
mediapipe==0.8.3.1
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
pathtools==0.1.2
Pillow==8.3.2
promise==2.3
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynput==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-xlib==0.29
pytorch-lightning==1.2.10
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.8.1
torchmetrics==0.2.0
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.10.29
Werkzeug==1.0.1
yarl==1.6.3
